bugfix: don't push a config if nothing changed
BaseController in clusters.py keeps track of the latest config so
downstream controllers can access it. When resolving a resource,
it always set the config to the latest config. However,
clusters.thatone.ai expects the config to be None if there are no
changes.

There are two possible fixes: one is to keep track of the old config
and compare it against the final config. The other is to keep track
of whether any operation gets resolved with the new config and only
then send an updated config. The latter seems to align better with
how people would expect the API to behave.